Docker Compose Executor boostrap
