+++ Dockerfile.test
@@ -4,6 +4,12 @@ ADD build/agent /app/agent
 ADD server.key /app/server.key
 ADD server.crt /app/server.crt
 
+RUN apt-get update && \
+  apt-get install -y python python-dev python-distribute python-pip && \
+  apt-get install curl -y && \
+  curl -sSL https://get.docker.com/ | sh && \
+  pip install docker-compose
+
 WORKDIR /app
 
 CMD ./agent serve
