+++ pkg/executors/docker_compose_executor.go
@@ -1,4 +1,4 @@
-package shell
+package executors
 
 import (
 	"bufio"
@@ -11,42 +11,59 @@ import (
 	"os/exec"
 	"path"
 	"regexp"
-	"runtime"
 	"strconv"
 	"strings"
 	"time"
 
 	pty "github.com/kr/pty"
 	api "github.com/semaphoreci/agent/pkg/api"
-	executors "github.com/semaphoreci/agent/pkg/executors"
 )
 
-type ShellExecutor struct {
-	executors.Executor
+type DockerComposeExecutor struct {
+	Executor
 
-	eventHandler  *executors.EventHandler
-	terminal      *exec.Cmd
-	tty           *os.File
-	stdin         io.Writer
-	stdoutScanner *bufio.Scanner
+	tmpDirectory              string
+	dockerConfiguration       api.Compose
+	dockerComposeManifestPath string
+	eventHandler              *EventHandler
+	terminal                  *exec.Cmd
+	tty                       *os.File
+	stdin                     io.Writer
+	stdoutScanner             *bufio.Scanner
 }
 
-func NewShellExecutor() *ShellExecutor {
-	return &ShellExecutor{}
+func NewDockerComposeExecutor(dockerConfiguration api.Compose) *DockerComposeExecutor {
+	return &DockerComposeExecutor{
+		dockerConfiguration:       dockerConfiguration,
+		dockerComposeManifestPath: "/tmp/docker-compose.yml",
+		tmpDirectory:              "/tmp/111", // make a better random name
+	}
 }
 
-func (e *ShellExecutor) Prepare() int {
-	e.terminal = exec.Command("bash", "--login")
+func (e *DockerComposeExecutor) Prepare() int {
+	compose := ConstructDockerComposeFile(e.dockerConfiguration)
+	ioutil.WriteFile(e.dockerComposeManifestPath, []byte(compose), 0644)
+
+	e.terminal = exec.Command(
+		"docker-compose",
+		"-f",
+		e.dockerComposeManifestPath,
+		"run",
+		"-v",
+		fmt.Sprintf("%s:%s:ro", e.tmpDirectory, e.tmpDirectory),
+		"main",
+		"bash",
+	)
 
 	return 0
 }
 
-func (e *ShellExecutor) Start() int {
+func (e *DockerComposeExecutor) Start() int {
 	log.Printf("[SHELL] Starting stateful shell")
 
 	tty, err := pty.Start(e.terminal)
 	if err != nil {
-		log.Printf("[SHELL] Failed to start stateful shell")
+		log.Printf("[SHELL] Failed to start stateful shell err: %+v", err)
 		return 1
 	}
 
@@ -60,7 +77,7 @@ func (e *ShellExecutor) Start() int {
 	return 0
 }
 
-func (e *ShellExecutor) silencePromptAndDisablePS1() {
+func (e *DockerComposeExecutor) silencePromptAndDisablePS1() {
 	everythingIsReadyMark := "87d140552e404df69f6472729d2b2c3"
 
 	e.stdin.Write([]byte("export PS1=''\n"))
@@ -100,17 +117,17 @@ func (e *ShellExecutor) silencePromptAndDisablePS1() {
 	log.Println("[SHELL] Initialization complete")
 }
 
-func (e *ShellExecutor) ExportEnvVars(envVars []api.EnvVar, callback executors.EventHandler) int {
+func (e *DockerComposeExecutor) ExportEnvVars(envVars []api.EnvVar, callback EventHandler) int {
 	commandStartedAt := int(time.Now().Unix())
 	directive := fmt.Sprintf("Exporting environment variables")
 	exitCode := 0
 
-	callback(executors.NewCommandStartedEvent(directive))
+	callback(NewCommandStartedEvent(directive))
 
 	defer func() {
 		commandFinishedAt := int(time.Now().Unix())
 
-		callback(executors.NewCommandFinishedEvent(
+		callback(NewCommandFinishedEvent(
 			directive,
 			exitCode,
 			commandStartedAt,
@@ -121,7 +138,7 @@ func (e *ShellExecutor) ExportEnvVars(envVars []api.EnvVar, callback executors.E
 	envFile := ""
 
 	for _, e := range envVars {
-		callback(executors.NewCommandOutputEvent(fmt.Sprintf("Exporting %s\n", e.Name)))
+		callback(NewCommandOutputEvent(fmt.Sprintf("Exporting %s\n", e.Name)))
 
 		value, err := base64.StdEncoding.DecodeString(e.Value)
 
@@ -133,19 +150,22 @@ func (e *ShellExecutor) ExportEnvVars(envVars []api.EnvVar, callback executors.E
 		envFile += fmt.Sprintf("export %s=%s\n", e.Name, ShellQuote(string(value)))
 	}
 
-	err := ioutil.WriteFile("/tmp/.env", []byte(envFile), 0644)
+	envPath := fmt.Sprintf("%s/.env", e.tmpDirectory)
+	err := ioutil.WriteFile(envPath, []byte(envFile), 0644)
 
 	if err != nil {
-		exitCode = 1
+		exitCode = 255
 		return exitCode
 	}
 
-	exitCode = e.RunCommand("source /tmp/.env", executors.DevNullEventHandler)
+	cmd := fmt.Sprintf("source %s", envPath)
+	exitCode = e.RunCommand(cmd, DevNullEventHandler)
 	if exitCode != 0 {
 		return exitCode
 	}
 
-	exitCode = e.RunCommand("echo 'source /tmp/.env' >> ~/.bash_profile", executors.DevNullEventHandler)
+	cmd = fmt.Sprintf("echo 'source %s' >> ~/.bash_profile", envPath)
+	exitCode = e.RunCommand(cmd, DevNullEventHandler)
 	if exitCode != 0 {
 		return exitCode
 	}
@@ -153,24 +173,33 @@ func (e *ShellExecutor) ExportEnvVars(envVars []api.EnvVar, callback executors.E
 	return exitCode
 }
 
-func (e *ShellExecutor) InjectFiles(files []api.File, callback executors.EventHandler) int {
+func (e *DockerComposeExecutor) InjectFiles(files []api.File, callback EventHandler) int {
 	directive := fmt.Sprintf("Injecting Files")
 	commandStartedAt := int(time.Now().Unix())
 	exitCode := 0
 
-	callback(executors.NewCommandStartedEvent(directive))
+	callback(NewCommandStartedEvent(directive))
 
 	for _, f := range files {
 		output := fmt.Sprintf("Injecting %s with file mode %s\n", f.Path, f.Mode)
 
-		callback(executors.NewCommandOutputEvent(output))
+		callback(NewCommandOutputEvent(output))
 
 		content, err := base64.StdEncoding.DecodeString(f.Content)
 
 		if err != nil {
-			callback(executors.NewCommandOutputEvent("Failed to decode content of file.\n"))
+			callback(NewCommandOutputEvent("Failed to decode content of file.\n"))
 			exitCode = 1
 			return exitCode
+		}
+
+		tmpPath := fmt.Sprintf("%s/file", e.tmpDirectory)
+
+		err = ioutil.WriteFile(tmpPath, []byte(content), 0644)
+		if err != nil {
+			callback(NewCommandOutputEvent(err.Error() + "\n"))
+			exitCode = 255
+			break
 		}
 
 		destPath := ""
@@ -181,32 +210,34 @@ func (e *ShellExecutor) InjectFiles(files []api.File, callback executors.EventHa
 			destPath = UserHomeDir() + "/" + f.Path
 		}
 
-		err = os.MkdirAll(path.Dir(destPath), os.ModePerm)
-		if err != nil {
-			callback(executors.NewCommandOutputEvent(err.Error() + "\n"))
-			exitCode = 1
+		cmd := fmt.Sprintf("mkdir -p %s", path.Dir(destPath))
+		exitCode = e.RunCommand(cmd, DevNullEventHandler)
+		if exitCode != 0 {
+			output := fmt.Sprintf("Failed to create destination path %s", destPath)
+			callback(NewCommandOutputEvent(output + "\n"))
 			break
 		}
 
-		err = ioutil.WriteFile(destPath, []byte(content), 0644)
-		if err != nil {
-			callback(executors.NewCommandOutputEvent(err.Error() + "\n"))
-			exitCode = 1
+		cmd = fmt.Sprintf("cp %s %s", tmpPath, destPath)
+		exitCode = e.RunCommand(cmd, DevNullEventHandler)
+		if exitCode != 0 {
+			output := fmt.Sprintf("Failed to move to destination path %s %s", tmpPath, destPath)
+			callback(NewCommandOutputEvent(output + "\n"))
 			break
 		}
 
-		cmd := fmt.Sprintf("chmod %s %s", f.Mode, destPath)
-		exitCode = e.RunCommand(cmd, executors.DevNullEventHandler)
+		cmd = fmt.Sprintf("chmod %s %s", f.Mode, destPath)
+		exitCode = e.RunCommand(cmd, DevNullEventHandler)
 		if exitCode != 0 {
 			output := fmt.Sprintf("Failed to set file mode to %s", f.Mode)
-			callback(executors.NewCommandOutputEvent(output + "\n"))
+			callback(NewCommandOutputEvent(output + "\n"))
 			break
 		}
 	}
 
 	commandFinishedAt := int(time.Now().Unix())
 
-	callback(executors.NewCommandFinishedEvent(
+	callback(NewCommandFinishedEvent(
 		directive,
 		exitCode,
 		commandStartedAt,
@@ -216,12 +247,12 @@ func (e *ShellExecutor) InjectFiles(files []api.File, callback executors.EventHa
 	return exitCode
 }
 
-func (e *ShellExecutor) RunCommand(command string, callback executors.EventHandler) int {
+func (e *DockerComposeExecutor) RunCommand(command string, callback EventHandler) int {
 	var err error
 
 	log.Printf("[SHELL] Running command: %s", command)
 
-	cmdFilePath := "/tmp/current-agent-cmd"
+	cmdFilePath := fmt.Sprintf("%s/current-agent-cmd", e.tmpDirectory)
 	startMark := "87d140552e404df69f6472729d2b2c1"
 	finishMark := "97d140552e404df69f6472729d2b2c2"
 
@@ -231,7 +262,7 @@ func (e *ShellExecutor) RunCommand(command string, callback executors.EventHandl
 	defer func() {
 		commandFinishedAt := int(time.Now().Unix())
 
-		callback(executors.NewCommandFinishedEvent(
+		callback(NewCommandFinishedEvent(
 			command,
 			exitCode,
 			commandStartedAt,
@@ -246,7 +277,14 @@ func (e *ShellExecutor) RunCommand(command string, callback executors.EventHandl
 	// Multiline commands don't work very well with the start/finish marker scheme.
 	// To circumvent this, we are storing the command in a file
 	//
-	ioutil.WriteFile(cmdFilePath, []byte(command), 0644)
+	err = ioutil.WriteFile(cmdFilePath, []byte(command), 0644)
+
+	if err != nil {
+		callback(NewCommandStartedEvent(command))
+		callback(NewCommandOutputEvent(fmt.Sprintf("Failed to run command: %+v\n", err)))
+
+		return 1
+	}
 
 	// Constructing command with start and end markers:
 	//
@@ -267,67 +305,50 @@ func (e *ShellExecutor) RunCommand(command string, callback executors.EventHandl
 
 	e.stdin.Write([]byte(commandWithStartAndEndMarkers))
 
+	stdoutScanner := bufio.NewScanner(e.tty)
+
 	log.Println("[SHELL] Scan started")
+	for stdoutScanner.Scan() {
+		t := stdoutScanner.Text()
+		log.Printf("[SHELL] (tty) %s\n", t)
 
-	err = executors.ScanLines(e.tty, func(line string) bool {
-		log.Printf("[SHELL] (tty) %s\n", line)
-
-		if strings.Contains(line, startMark) {
-			log.Printf("[SHELL] Detected command start")
+		if strings.Contains(t, startMark) {
 			streamEvents = true
 
-			callback(executors.NewCommandStartedEvent(command))
+			callback(NewCommandStartedEvent(command))
 
-			return true
+			continue
 		}
 
-		if strings.Contains(line, finishMark) {
-			log.Printf("[SHELL] Detected command end")
-
-			finalOutputPart := strings.Split(line, finishMark)
-
-			// if there is anything else other than the command end marker
-			// print it to the user
-			if finalOutputPart[0] != "" {
-				callback(executors.NewCommandOutputEvent(finalOutputPart[0] + "\n"))
-			}
-
+		if strings.Contains(t, finishMark) {
 			streamEvents = false
 
-			if match := commandEndRegex.FindStringSubmatch(line); len(match) == 2 {
-				log.Printf("[SHELL] Parsing exit status succedded")
-
+			if match := commandEndRegex.FindStringSubmatch(t); len(match) == 2 {
 				exitCode, err = strconv.Atoi(match[1])
 
 				if err != nil {
 					log.Printf("[SHELL] Panic while parsing exit status, err: %+v", err)
 
-					callback(executors.NewCommandOutputEvent("Failed to read command exit code\n"))
+					callback(NewCommandOutputEvent("Failed to read command exit code\n"))
 				}
 
-				log.Printf("[SHELL] Setting exit code to %d", exitCode)
 			} else {
-				log.Printf("[SHELL] Failed to parse exit status")
-
 				exitCode = 1
-				callback(executors.NewCommandOutputEvent("Failed to read command exit code\n"))
+				callback(NewCommandOutputEvent("Failed to read command exit code\n"))
 			}
 
-			log.Printf("[SHELL] Stopping scanner")
-			return false
+			break
 		}
 
 		if streamEvents {
-			callback(executors.NewCommandOutputEvent(line + "\n"))
+			callback(NewCommandOutputEvent(t + "\n"))
 		}
-
-		return true
-	})
+	}
 
 	return exitCode
 }
 
-func (e *ShellExecutor) Stop() int {
+func (e *DockerComposeExecutor) Stop() int {
 	err := e.terminal.Process.Kill()
 
 	if err != nil {
@@ -337,30 +358,6 @@ func (e *ShellExecutor) Stop() int {
 	return 0
 }
 
-func (e *ShellExecutor) Cleanup() int {
+func (e *DockerComposeExecutor) Cleanup() int {
 	return 0
-}
-
-func UserHomeDir() string {
-	if runtime.GOOS == "windows" {
-		home := os.Getenv("HOMEDRIVE") + os.Getenv("HOMEPATH")
-		if home == "" {
-			home = os.Getenv("USERPROFILE")
-		}
-		return home
-	}
-	return os.Getenv("HOME")
-}
-
-func ShellQuote(s string) string {
-	pattern := regexp.MustCompile(`[^\w@%+=:,./-]`)
-
-	if len(s) == 0 {
-		return "''"
-	}
-	if pattern.MatchString(s) {
-		return "'" + strings.Replace(s, "'", "'\"'\"'", -1) + "'"
-	}
-
-	return s
 }
