+++ pkg/executors/executor.go
@@ -1,6 +1,8 @@
 package executors
 
 import (
+	"fmt"
+
 	api "github.com/semaphoreci/agent/pkg/api"
 )
 
@@ -12,4 +14,18 @@ type Executor interface {
 	RunCommand(string, EventHandler) int
 	Stop() int
 	Cleanup() int
+}
+
+const ExecutorTypeShell = "shell"
+const ExecutorTypeDockerCompose = "dockercompose"
+
+func CreateExecutor(request *api.JobRequest) (Executor, error) {
+	switch request.Executor {
+	case ExecutorTypeShell:
+		return NewShellExecutor(), nil
+	case ExecutorTypeDockerCompose:
+		return NewDockerComposeExecutor(request.Compose), nil
+	default:
+		return nil, fmt.Errorf("Uknown executor type")
+	}
 }
