+++ test/e2e/docker_hello_world.rb
@@ -7,15 +7,24 @@ start_job <<-JSON
   {
     "id": "#{$JOB_ID}",
 
-    "executor": "shell",
+    "executor": "dockercompose",
+
+    "compose": {
+      "containers": [
+        {
+          "name": "main",
+          "image" "ruby:2.6"
+        }
+      ]
+    },
 
     "env_vars": [],
 
     "files": [],
 
     "commands": [
-      { "directive": "sleep 10" },
-      { "directive": "echo 'here'" }
+      { "directive": "echo Hello World" },
+      { "directive": "ruby --version" }
     ],
 
     "epilogue_commands": [],
@@ -28,11 +37,7 @@ start_job <<-JSON
   }
 JSON
 
-sleep 1
-
-stop_job
-
-sleep 1
+wait_for_job_to_finish
 
 assert_job_log <<-LOG
   {"event":"job_started",  "timestamp":"*"}
@@ -40,7 +45,10 @@ assert_job_log <<-LOG
   {"event":"cmd_finished", "timestamp":"*", "directive":"Exporting environment variables","exit_code":0,"finished_at":"*","started_at":"*"}
   {"event":"cmd_started",  "timestamp":"*", "directive":"Injecting Files"}
   {"event":"cmd_finished", "timestamp":"*", "directive":"Injecting Files","exit_code":0,"finished_at":"*","started_at":"*"}
-  {"event":"cmd_started",  "timestamp":"*", "directive":"sleep 10"}
-  {"event":"cmd_finished", "timestamp":"*", "directive":"sleep 10","exit_code":1,"finished_at":"*","started_at":"*"}
-  {"event":"job_finished", "timestamp":"*", "result":"failed"}
+  {"event":"cmd_started",  "timestamp":"*", "directive":"echo Hello World"}
+  {"event":"cmd_output",   "timestamp":"*", "output":"Hello World\\n"}
+  {"event":"cmd_finished", "timestamp":"*", "directive":"echo Hello World","exit_code":0,"finished_at":"*","started_at":"*"}
+  {"event":"cmd_started",  "timestamp":"*", "directive":"export SEMAPHORE_JOB_RESULT=passed"}
+  {"event":"cmd_finished", "timestamp":"*", "directive":"export SEMAPHORE_JOB_RESULT=passed","exit_code":0,"finished_at":"*","started_at":"*"}
+  {"event":"job_finished", "timestamp":"*", "result":"passed"}
 LOG
