+++ test/e2e/env_vars.rb
@@ -7,6 +7,8 @@ start_job <<-JSON
   {
     "id": "#{$JOB_ID}",
 
+    "executor": "shell",
+
     "env_vars": [
       { "name": "A", "value": "#{`echo "hello" | base64`}" },
       { "name": "B", "value": "#{`echo "how are you?" | base64`}" },
