+++ test/e2e/file_injection_broken_file_mode.rb
@@ -7,6 +7,8 @@ start_job <<-JSON
   {
     "id": "#{$JOB_ID}",
 
+    "executor": "shell",
+
     "env_vars": [],
 
     "files": [
